requirements.txt
scikit-learn>=1.3.1
pandas>=2.1.1
matplotlib>=3.8.0
seaborn>=0.12.2
fastparquet
pymongo>=4.5.0
python-dotenv
pyyaml>=6.0.1
matplotlib>=3.7.3
onnx==1.14.1
#onnxruntime-gpu>=1.16.0 uncomment if you are using GPU
onnxruntime==1.16.0 # for cpu
loguru>=0.7.2
mlflow==2.7.1
evidently==0.4.8
pytorch-lightning==1.9.5
torchvision==0.15.2
tensorboard
-e .
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
# Pytorch data processing idea